quick

+++ database/migrations/2024_09_09_131254_add_to_users_company_column.php
@@ -12,6 +12,7 @@ return new class () extends Migration {
     {
         Schema::table('users', function (Blueprint $table) {
             $table->foreignUuid('company_id')->nullable()->constrained('companies')->nullOnDelete();
+            $table->boolean('responsible')->nullable()->default(false);
         });
     }
 
@@ -21,7 +22,15 @@ return new class () extends Migration {
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('company_id');
+            // Verificar e remover a chave estrangeira associada a company_id
+            if (Schema::hasColumn('users', 'company_id')) {
+                $table->dropForeign(['company_id']); // Remove a chave estrangeira
+                $table->dropColumn('company_id');    // Depois, remove a coluna
+            }
+
+            if (Schema::hasColumn('users', 'responsible')) {
+                $table->dropColumn('responsible');   // Remove a coluna responsible se ela existir
+            }
         });
     }
 };
